# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * How thorough the report should be
 */
export type ResearchInputDepth = typeof ResearchInputDepth[keyof typeof ResearchInputDepth];


export const ResearchInputDepth = {
  overview: 'overview',
  standard: 'standard',
  deep: 'deep',
} as const;

export interface ResearchInput {
  /** @minLength 1 */
  topic: string;
  /** How thorough the report should be */
  depth?: ResearchInputDepth;
}

export interface ResearchSection {
  heading: string;
  content: string;
}

export interface ResearchSource {
  title: string;
  note: string;
}

export interface ResearchReport {
  title: string;
  summary: string;
  sections: ResearchSection[];
  keyTakeaways: string[];
  furtherQuestions: string[];
  suggestedSources?: ResearchSource[];
}

export type EssayInputEssayType = typeof EssayInputEssayType[keyof typeof EssayInputEssayType];


export const EssayInputEssayType = {
  argumentative: 'argumentative',
  narrative: 'narrative',
  expository: 'expository',
  persuasive: 'persuasive',
  descriptive: 'descriptive',
  'compare-contrast': 'compare-contrast',
} as const;

export interface EssayInput {
  /** @minLength 1 */
  topic: string;
  essayType: EssayInputEssayType;
  /**
     * @minimum 100
     * @maximum 3000
     */
  wordCount: number;
  tone?: string;
}

export interface Essay {
  title: string;
  outline: string[];
  body: string;
  wordCount: number;
}

export type QuizInputDifficulty = typeof QuizInputDifficulty[keyof typeof QuizInputDifficulty];


export const QuizInputDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface QuizInput {
  /** @minLength 1 */
  topic: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  numQuestions: number;
  difficulty: QuizInputDifficulty;
}

export interface QuizQuestion {
  question: string;
  options: string[];
  correctIndex: number;
  explanation: string;
}

export interface Quiz {
  title: string;
  questions: QuizQuestion[];
}

export interface FlashcardSetInput {
  /** @minLength 1 */
  topic: string;
  /**
     * @minimum 1
     * @maximum 30
     */
  numCards: number;
}

export interface Flashcard {
  front: string;
  back: string;
}

export interface FlashcardSet {
  title: string;
  cards: Flashcard[];
}

export interface StudyNotesInput {
  /** @minLength 1 */
  topic: string;
}

export interface KeyTerm {
  term: string;
  definition: string;
}

export interface NoteSection {
  heading: string;
  bullets: string[];
}

export interface StudyNotes {
  title: string;
  sections: NoteSection[];
  keyTerms: KeyTerm[];
}

export interface PresentationInput {
  /** @minLength 1 */
  topic: string;
  /**
     * @minimum 3
     * @maximum 20
     */
  numSlides: number;
}

export interface Slide {
  title: string;
  bullets: string[];
  speakerNotes: string;
}

export interface Presentation {
  title: string;
  slides: Slide[];
}

export interface MathSolverInput {
  /**
     * The math problem to solve, e.g. "2x + 5 = 17" or "derivative of 3x^2 + 2x"
     * @minLength 1
     */
  problem: string;
}

export interface MathStep {
  /** Short label for this step, e.g. "Isolate the variable" */
  title: string;
  /** Plain-language explanation of what is happening in this step and why */
  explanation: string;
  /** The math expression/equation resulting from this step, as plain text (e.g. "2x = 12") */
  expression: string;
}

export interface MathSolution {
  restatedProblem: string;
  /** Short label for the branch of math, e.g. "Linear Equations", "Calculus - Derivatives" */
  topic: string;
  steps: MathStep[];
  finalAnswer: string;
  /** A brief note on how to verify/check the final answer */
  checkNote?: string;
}

export type TextTransformInputMode = typeof TextTransformInputMode[keyof typeof TextTransformInputMode];


export const TextTransformInputMode = {
  summarize: 'summarize',
  expand: 'expand',
  simplify: 'simplify',
  rewrite: 'rewrite',
  translate: 'translate',
  'fix-grammar': 'fix-grammar',
} as const;

export interface TextTransformInput {
  /** @minLength 1 */
  text: string;
  mode: TextTransformInputMode;
  targetLanguage?: string;
}

export interface TextTransformResult {
  result: string;
}