# lib/api-zod/src/generated/types/essayInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EssayInputEssayType } from './essayInputEssayType';

export interface EssayInput {
  /** @minLength 1 */
  topic: string;
  essayType: EssayInputEssayType;
  /**
     * @minimum 100
     * @maximum 3000
     */
  wordCount: number;
  tone?: string;
}