# lib/api-zod/src/generated/types/flashcardSet.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Flashcard } from './flashcard';

export interface FlashcardSet {
  title: string;
  cards: Flashcard[];
}