# lib/api-zod/src/generated/types/flashcardSetInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FlashcardSetInput {
  /** @minLength 1 */
  topic: string;
  /**
     * @minimum 1
     * @maximum 30
     */
  numCards: number;
}