# lib/api-zod/src/generated/types/mathSolverInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MathSolverInput {
  /**
     * The math problem to solve, e.g. "2x + 5 = 17" or "derivative of 3x^2 + 2x"
     * @minLength 1
     */
  problem: string;
}