# lib/api-zod/src/generated/types/mathStep.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MathStep {
  /** Short label for this step, e.g. "Isolate the variable" */
  title: string;
  /** Plain-language explanation of what is happening in this step and why */
  explanation: string;
  /** The math expression/equation resulting from this step, as plain text (e.g. "2x = 12") */
  expression: string;
}