# lib/api-zod/src/generated/types/presentation.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Slide } from './slide';

export interface Presentation {
  title: string;
  slides: Slide[];
}