# lib/api-zod/src/generated/types/presentationInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PresentationInput {
  /** @minLength 1 */
  topic: string;
  /**
     * @minimum 3
     * @maximum 20
     */
  numSlides: number;
}