# lib/api-zod/src/generated/types/quizInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuizInputDifficulty } from './quizInputDifficulty';

export interface QuizInput {
  /** @minLength 1 */
  topic: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  numQuestions: number;
  difficulty: QuizInputDifficulty;
}