# lib/api-zod/src/generated/types/researchInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResearchInputDepth } from './researchInputDepth';

export interface ResearchInput {
  /** @minLength 1 */
  topic: string;
  /** How thorough the report should be */
  depth?: ResearchInputDepth;
}