# lib/api-zod/src/generated/types/textTransformInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TextTransformInputMode } from './textTransformInputMode';

export interface TextTransformInput {
  /** @minLength 1 */
  text: string;
  mode: TextTransformInputMode;
  targetLanguage?: string;
}